More tuning of razoring
Bugfix from Tom; cutNode was not defined in the first instance. Thanks!
ss->reduction was a nonfunctional item, replaced with (ss-1)->reduction in another way but that was not succesful. What I am trying now instead is futility pruning code from 'Scid Serpent'
which logically could work in ALL nodes razoring also. It is hard to detect but you should see it in time to depth going down. Sorry if it is more complex.