Threats by Kings
This is a change I wrote to Marco about a long time ago and he said he would like to be kept updated so it is about time I think it is tested. It is not very strong I fear but I tested this in combination with a search change. A bit hard to explain but mainly simulated annealing reasons to test them together. It is still possible *if* they pass together that they also work together (enhance each other), because this part changes king safety a bit and that might help search, but might not scale. So my small scale testing anyway is not going to be enough so I hope I get permission to test it in the frame work. I will continue with some local testing.