Futility change
This part of the patch I have some higher hopes about, but no idea if it scales well. Reductions should scale well but, of course, they introduce also some risk. Because this is a change involving cutNode it works differently in SMP mode. It should work similar to the other cutNode reduction and the cutNode precondition is there for safety reasons (makes it more conservative).