More SEE based and 'improving' based futility pruning,
different LMR 'improving' based. Also this still has Tom's version of 'improving' mainly because this is what I tested also in version  I of branch. Same for enoughMaterial, here is a possible improvement to take it out because obsolete now with KBPK fix. It is difficult to make out what effect 'improving' numbers in qsearch have, maybe mainly a reduction of the qsearch futilitymargin Value(128) so it could be a bit risky and maybe one reduction (number) has the same effect (no quiet moves).