Small fix
Bishop_pair() was defined at a later date, but not yet integrated here. This increases the bench, but is logical change so maybe some retuning of bishop pair constant would be in order and measurable. Ryan's last try failed tuning the lineair coefficient,  but maybe a very small hike could do. (just a suggestion for further testing).